+++ test/unit/fuzz-tools.test.js
@@ -285,6 +285,7 @@ import * as crsV305 from '../../lib/crs-v305.js';
 import * as icansV306 from '../../lib/icans-v306.js';
 import * as dmeV307 from '../../lib/dme-v307.js';
 import * as concussionRtlV308 from '../../lib/concussion-rtl-v308.js';
+import * as gvhdV309 from '../../lib/gvhd-v309.js';
 import * as tbTesting from '../../lib/tb-testing.js';
 
 const MODULES = {
@@ -515,6 +516,7 @@ const MODULES = {
   'icans-v306.js': icansV306,
   'dme-v307.js': dmeV307,
   'concussion-rtl-v308.js': concussionRtlV308,
+  'gvhd-v309.js': gvhdV309,
   'tb-testing.js': tbTesting,
 };